+++ src/main/java/com/deviget/minesweeper/domain/model/game/Board.java
@@ -1,5 +1,9 @@
 package com.deviget.minesweeper.domain.model.game;
 
+import com.deviget.minesweeper.domain.model.game.cell.AdjacentCellCoordinates;
+import com.deviget.minesweeper.domain.model.game.cell.Cell;
+import com.deviget.minesweeper.domain.model.game.cell.CellCoordinate;
+
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
@@ -9,6 +13,7 @@ import java.util.stream.IntStream;
 
 public class Board {
 
+  private final BoardSize boardSize;
   private final List<Row> rows = new ArrayList<>();
 
   static Board empty(final Preferences preferences) {
@@ -22,14 +27,37 @@ public class Board {
                     .collect(Collectors.toList());
 
     this.rows.addAll(rowsWithEmptyCells);
+    this.boardSize = preferences.boardSize();
   }
 
-  void redefineCells(final CellCoordinate firstUncoveredCell, final Set<CellCoordinate> mineCoordinates) {
+  void placeMines(final Set<CellCoordinate> mineCoordinates) {
+    mineCoordinates.forEach(mineCoordinate -> {
+      rowAt(mineCoordinate.rowIndex()).placeMine(mineCoordinate.cellIndex());
+      this.incrementAdjacentMineAlert(mineCoordinate);
+    });
+  }
 
+  private void incrementAdjacentMineAlert(final CellCoordinate mineCoordinate) {
+    AdjacentCellCoordinates.resolve(boardSize, mineCoordinate).forEach(this::incrementMineAlert);
+  }
+
+  private void incrementMineAlert(final CellCoordinate cellCoordinate) {
+    rowAt(cellCoordinate.rowIndex()).incrementMineAlert(cellCoordinate.cellIndex());
+  }
+
+  Row rowAt(final int rowIndex) {
+    return this.rows.get(rowIndex);
   }
 
   List<Row> rows() {
     return Collections.unmodifiableList(rows);
   }
 
+  Cell uncoverCell(final CellCoordinate cellCoordinate) {
+    return Cell.emptyAt(0);
+  }
+
+  boolean hasOnlyCoveredMineCells() {
+    return rows.stream().allMatch(row -> row.cells().stream().allMatch(cell -> !cell.isUncovered() && cell.isMine()));
+  }
 }