+++ src/main/java/com/deviget/minesweeper/domain/model/game/Game.java
@@ -1,18 +1,21 @@
 package com.deviget.minesweeper.domain.model.game;
 
 import com.deviget.minesweeper.domain.model.UserId;
+import com.deviget.minesweeper.domain.model.game.cell.CellCoordinate;
 
 import java.time.Instant;
+import java.util.HashSet;
 import java.util.List;
+import java.util.Set;
 
 public class Game {
 
   private final GameId gameId;
-  private final Board board;
-  private final GameStatus status;
   private final UserId userId;
   private final Preferences preferences;
-  private long startedOn;
+  private final Board board;
+  private GameStatus status;
+  private final long startedOn;
   private long updatedOn;
 
   public static Game configure(final Preferences preferences,
@@ -26,9 +29,41 @@ public class Game {
     this.status = GameStatus.NEW;
     this.preferences = preferences;
     this.board = Board.empty(preferences);
-    this.userId = userId;
     this.startedOn = Instant.now().toEpochMilli();
     this.updatedOn = startedOn;
+    this.userId = userId;
+  }
+
+  public void uncoverCell(final CellCoordinate cellCoordinate) {
+    timestamp();
+    if(isNew()) {
+      this.status = GameStatus.ONGOING;
+      this.board.placeMines(randomizeMineCoordinates(cellCoordinate));
+    }
+    if(this.board.uncoverCell(cellCoordinate).isMine()) {
+      this.status = GameStatus.LOST;
+    } else if(this.board.hasOnlyCoveredMineCells()) {
+      this.status = GameStatus.WON;
+    }
+  }
+
+  private Set<CellCoordinate> randomizeMineCoordinates(final CellCoordinate ignoredCoordinate) {
+    final Set<CellCoordinate> mineCoordinates = new HashSet<>();
+
+    while(mineCoordinates.size() < preferences.numberOfMines()) {
+      final CellCoordinate randomCoordinate =
+              CellCoordinate.random(preferences.rows(), preferences.columns());
+
+      if(!ignoredCoordinate.equals(randomCoordinate)) {
+        mineCoordinates.add(randomCoordinate);
+      }
+    }
+
+    return mineCoordinates;
+  }
+
+  private void timestamp() {
+    this.updatedOn = Instant.now().toEpochMilli();
   }
 
   public GameId id() {
@@ -53,6 +88,14 @@ public class Game {
 
   public long updatedOn() {
     return updatedOn;
+  }
+
+  boolean isNew() {
+    return this.status.equals(GameStatus.NEW);
+  }
+
+  boolean isOver() {
+    return this.status.equals(GameStatus.WON) || this.status.equals(GameStatus.LOST);
   }
 
 }