+++ src/main/java/com/deviget/minesweeper/domain/model/game/MineAlertCell.java
@@ -1,9 +1,0 @@
-package com.deviget.minesweeper.domain.model.game;
-
-public class MineAlertCell extends Cell {
-
-  protected MineAlertCell(final int index) {
-    super(index, CellType.MINE_ALERT);
-  }
-
-}