+++ src/main/java/com/deviget/minesweeper/domain/model/game/Preferences.java
@@ -44,4 +44,7 @@ public class Preferences {
     return numberOfMines;
   }
 
+  public BoardSize boardSize() {
+    return new BoardSize(rows, columns);
+  }
 }