+++ src/main/java/com/deviget/minesweeper/domain/model/game/Row.java
@@ -1,15 +1,17 @@
 package com.deviget.minesweeper.domain.model.game;
 
+import com.deviget.minesweeper.domain.model.game.cell.Cell;
+
 import java.util.Collections;
-import java.util.LinkedHashSet;
 import java.util.Set;
+import java.util.TreeSet;
 import java.util.stream.Collectors;
 import java.util.stream.IntStream;
 
 public class Row {
 
   private final int index;
-  private final Set<Cell> cells = new LinkedHashSet<>();
+  private final Set<Cell> cells = new TreeSet<>();
 
   public static Row empty(final int index, final int numberOfCells) {
     return new Row(index, numberOfCells);
@@ -20,16 +22,25 @@ public class Row {
     this.cells.addAll(IntStream.range(0, numberOfCells).mapToObj(Cell::emptyAt).collect(Collectors.toSet()));
   }
 
-  public void addMineAt(final Board board, final int cellIndex) {
-
+  public void placeMine(final int cellIndex) {
+    replaceCell(Cell.mineAt(cellIndex));
   }
 
-  public void incrementMineAlertAt(final int cellIndex) {
+  public void incrementMineAlert(final int cellIndex) {
+    replaceCell(cellAt(cellIndex).incrementMine());
+  }
 
+  private void replaceCell(final Cell cell) {
+    this.cells.remove(cell);
+    this.cells.add(cell);
   }
 
   public Set<Cell> cells() {
     return Collections.unmodifiableSet(cells);
   }
 
+  Cell cellAt(final int cellIndex) {
+    return cells.stream().filter(c -> c.matchIndex(cellIndex)).findFirst()
+            .orElseThrow(() -> new IllegalArgumentException("Unable to find cell with index " + cellIndex));
+  }
 }