+++ src/main/java/com/deviget/minesweeper/domain/model/game/cell/CellNavigationResolver.java
@@ -1,0 +1,9 @@
+package com.deviget.minesweeper.domain.model.game.cell;
+
+import com.deviget.minesweeper.domain.model.game.BoardSize;
+
+public interface CellNavigationResolver {
+
+  CellCoordinate resolve(final BoardSize boardSize, final CellCoordinate currentCoordinate);
+
+}