+++ src/main/java/com/deviget/minesweeper/domain/model/game/cell/LowNavigationResolver.java
@@ -1,0 +1,16 @@
+package com.deviget.minesweeper.domain.model.game.cell;
+
+import com.deviget.minesweeper.domain.model.game.BoardSize;
+
+public class LowNavigationResolver implements CellNavigationResolver {
+
+  @Override
+  public CellCoordinate resolve(final BoardSize boardSize,
+                                final CellCoordinate currentCoordinate) {
+    if(currentCoordinate.isBottom(boardSize)) {
+      return CellCoordinate.invalid();
+    }
+    return CellCoordinate.with(currentCoordinate.rowIndex + 1, currentCoordinate.cellIndex);
+  }
+
+}