+++ src/main/java/com/deviget/minesweeper/domain/model/game/cell/RightNavigationResolver.java
@@ -1,0 +1,14 @@
+package com.deviget.minesweeper.domain.model.game.cell;
+
+import com.deviget.minesweeper.domain.model.game.BoardSize;
+
+public class RightNavigationResolver implements CellNavigationResolver {
+  @Override
+  public CellCoordinate resolve(final BoardSize boardSize,
+                                final CellCoordinate currentCoordinate) {
+    if(currentCoordinate.isRightEnd(boardSize)) {
+      return CellCoordinate.invalid();
+    }
+    return CellCoordinate.with(currentCoordinate.rowIndex, currentCoordinate.cellIndex + 1);
+  }
+}